organized and updated namespaces of inspectors
