+++ XRTK-Core/Assets/XRTK.Tests/InputSystem/TestFixture_02_InteractionDefinitionTests.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System;
 using NUnit.Framework;
 using System.Diagnostics;
 using UnityEngine;
