+++ XRTK-Core/Packages/com.xrtk.core/Extensions/CameraExtensions.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEngine;
-using XRTK.Utilities;
 
 namespace XRTK.Extensions
 {
