+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/InputSystem.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 28e06a5d68024a0d8e7fcc4368e1a4c8
+guid: f6de311ecdb5d8a4eb7b184d6e3bf303
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
