+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/InputSystem/Controllers/OpenVR/OculusGoControllerMappingProfileInspector.cs
@@ -2,9 +2,9 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEditor;
-using XRTK.Inspectors.Profiles;
+using XRTK.Definitions.Controllers.OpenVR;
 
-namespace XRTK.Definitions.Controllers.OpenVR.Inspectors.Profiles
+namespace XRTK.Inspectors.Profiles.InputSystem.Controllers.OpenVR
 {
     [CustomEditor(typeof(OculusGoControllerMappingProfile))]
     public class OculusGoControllerMappingProfileInspector : BaseMixedRealityControllerMappingProfileInspector { }
