+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/InputSystem/MixedRealityControllerDataProviderProfileInspector.cs
@@ -7,7 +7,7 @@ using XRTK.Inspectors.Utilities;
 using XRTK.Definitions.Controllers;
 using XRTK.Services;
 
-namespace XRTK.Inspectors.Profiles
+namespace XRTK.Inspectors.Profiles.InputSystem
 {
     [CustomEditor(typeof(MixedRealityControllerDataProvidersProfile))]
     public class MixedRealityControllerDataProviderProfileInspector : BaseMixedRealityProfileInspector
