+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/SpatialAwarenessSystem/BaseMixedRealitySpatialObserverProfileInspector.cs
